translation-agent/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
anthropic==0.96.0
jieba==0.42.1
pydantic==2.10.3
aiofiles==24.1.0
python-multipart==0.0.20
pytest==8.3.4
certifi==2026.2.28
